Lapunik/requirements.txt
pygame
mysql-connector-python
flask
